requirements.txt
pyrogram
beautifulsoup4
requests
python-dotenv
tgcrypto
